+++ src/components/chat/chatRoom/ChatRoomPanel.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState, useCallback } from 'react';
 import { useParams } from 'next/navigation';
 import Header from '@/components/common/Header';
 import ChatRoom from './ChatRoom';
-import ChatInputBar from '@/components/chat/chatRoom/ChatInputBar';
+import ChatInputBar from '@/components/chat/chatRoom/input/ChatInputBar';
 import { ChatMessagePropsType, ChatHeaderInfoPropsType } from '@/types/chat';
 import { getChatMessage, getChatHeaderInfo } from '@/api/chatApi';
 import { useChatSocket } from '@/hooks/chat/useChatSocket';
